requirements.txt
sentence-transformers
torch
transformers
numpy
fastapi
uvicorn
requests
python-dotenv
python-multipart
pydantic
mysql-connector-python
sqlalchemy
mysqlclient
pymysql
cryptography